Split file tool executor boundaries
